[hotfix] Make NoOpTransientBlobService a public testing class
